making new component for snippet
