+++ docs/app/pages/components/components-sections/icon/icon/icon.component.html
@@ -48,10 +48,11 @@
 <hr>
 
 <p>The <code>ux-icon</code> component can be used to dynamically display icons and allows easy control of
-    their size and orientation. Icons can be selected from the UX Aspects <a [routerLink]="['/css/icons']">
-    ux-icons </a>set. The <code>ux-icon</code> component also allows alternative icons to be loaded from a
-    custom icon set to override default icons. For example, icons for a specific project could be implemented
-    via a custom set, or an icon set created for a particular size could be used.</p>
+    their size and orientation. Icons can be selected from the UX Aspects <a fragment="ux-icons"
+    [routerLink]="[iconSetDocumentationRoute]">ux-icons </a>set. The <code>ux-icon</code> component also
+    allows alternative icons to be loaded from a custom icon set to override default icons. For example,
+    icons for a specific project could be implemented via a custom set, or an icon set created for a
+    particular size could be used.</p>
 <p>
     The following properties can be used to configure the component:
 </p>
