+++ docs/app/pages/components/components-sections/icon/icon/icon.component.ts
@@ -1,8 +1,9 @@
-import { Component } from '@angular/core';
+import {Component, Inject } from '@angular/core';
 import { BaseDocumentationSection } from '../../../../../components/base-documentation-section/base-documentation-section';
 import { DocumentationSectionComponent } from '../../../../../decorators/documentation-section-component';
 import { IPlayground } from '../../../../../interfaces/IPlayground';
 import { IPlaygroundProvider } from '../../../../../interfaces/IPlaygroundProvider';
+import { DOCUMENTATION_TOKEN, DocumentationType } from "../../../../../services/playground/tokens/documentation.token";
 
 @Component({
     selector: 'uxd-components-icon',
@@ -12,6 +13,8 @@ import { IPlaygroundProvider } from '../../../../../interfaces/IPlaygroundProvid
 
 @DocumentationSectionComponent('ComponentsIconComponent')
 export class ComponentsIconComponent extends BaseDocumentationSection implements IPlaygroundProvider {
+
+    iconSetDocumentationRoute: string;
 
     playground: IPlayground = {
         files: {
@@ -27,8 +30,10 @@ export class ComponentsIconComponent extends BaseDocumentationSection implements
         ]
     };
 
-    constructor() {
+    constructor(@Inject(DOCUMENTATION_TOKEN) private _documentationType: DocumentationType) {
         super(require.context('./snippets/', false, /(html|css|js|ts)$/));
+
+        this.iconSetDocumentationRoute = _documentationType === DocumentationType.MicroFocus ? '/ui-components/styling#ux-icons' : '/css/icons';
     }
 
 }
