+++ docs/app/pages/css/css-sections/icons/ux-icons/icon-preview/icon-snippet/icon-snippet.component.html
@@ -1,0 +1,10 @@
+<uxd-snippet [code]="code">
+</uxd-snippet>
+
+<button [uxTooltip]="tooltipTemplate" (click)="copy()" type="button" class="btn btn-link btn-icon button-secondary" aria-label="copy">
+    <ux-icon name="clipboard" aria-hidden="true"></ux-icon>
+</button>
+
+<ng-template #tooltipTemplate>
+    <div class="usage-tooltip">Copy to Clipboard</div>
+</ng-template>
