+++ docs/app/pages/css/css-sections/icons/ux-icons/ux-icons.component.html
@@ -1,7 +1,8 @@
-<p>To use an icon simply apply the <code>{{ iconClass }}</code> class along with one of the following
-    classes <ng-container *ngIf="iconset === 'ux-icons'">or use the <code>ux-icon</code><a
-    [routerLink]="['/components/icons']"> component</a> along with one of the following
-    names</ng-container>. Click on an icon to view its implementation.</p>
+<p *ngIf="iconClass === 'hpe-icon'">To use an icon simply apply the <code>{{ iconClass }}</code> class along
+    with one of the following classes:</p>
+<p  *ngIf="iconClass === 'ux-icon'">To use an icon apply the <code>ux-icon</code><a
+    [routerLink]="['/components/icons']"> component</a> along with one of the following names. Click on an
+    icon to view its implementation.</p>
 
 <div class="row m-b-md">
     <div class="col-xs-6">
@@ -29,7 +30,7 @@
             [name]="icon.name"
             [classname]="icon.classname"
             [iconset]="iconset"
-            [iconClass]="iconClass">
+            [iconSetClass]="iconClass">
         </uxd-icon-preview>
     </div>
 </div>
